Complete refactoring of the code. Everything is now separated into individual files to improve readability. This has been tested, it runs flawlessly.
