+++ src/temp/DictionaryGerman.py
