+++ src/user.py
@@ -1,0 +1,4 @@
+class User:
+    def __init__(self, name, face):
+        self.name = name
+        self.face = face
